Fixed handlers to allow debug

+++ BoatControlPanel/BaseBoatPage.h
@@ -41,7 +41,7 @@ struct FloatStateUpdate {
 };
 
 struct IntStateUpdate {
-    int value;
+    int16_t value;
 };
 
 struct BoolStateUpdate {
